chore(extension-examples): refactor to new modes and files
