--- packages/extension-example-1/src/html/client.ts
@@ -1,1 +1,0 @@
-export { default } from ".";
