--- packages/extension-example-1/src/html/server.ts
@@ -1,1 +1,0 @@
-export { default } from ".";
