--- packages/extension-example-2/src/amp/server.ts
@@ -1,1 +1,0 @@
-export { default } from "../html";
