+++ packages/extension-example-2/src/assets/frontity.png
