+++ packages/extension-example-2/src/client.ts
@@ -1,0 +1,1 @@
+export { default } from "..";
