+++ packages/extension-example-2/src/components/dynamic2.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { css } from "frontity";
-import frontity from "./frontity.png";
+import frontity from "../assets/frontity.png";
 
 const Dynamic = () => (
   <div
