--- packages/extension-example-2/src/html/client.ts
@@ -1,1 +1,0 @@
-export { default } from "./";
