+++ packages/extension-example-2/src/server.ts
@@ -1,5 +1,5 @@
-import Type from "../..";
-import ExtensionExample2 from "./";
+import Type from "..";
+import ExtensionExample2 from ".";
 
 const beforeSSR: Type["actions"]["theme"]["beforeSSR"] = ({ state }) => {
   state.theme.prop1 = 2;
